requirements.txt
git+https://github.com/tencent-ailab/IP-Adapter.git
diffusers
transformers
accelerate
scipy
safetensors
controlnet_aux
pillow
opencv-python
python-dotenv



